add[no-merged] adding checkbox, filter, create component for table, like tablehead, tablebody
