+++ package.json
@@ -11,11 +11,8 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@emotion/react": "^11.10.6",
-    "@emotion/styled": "^11.10.6",
     "@fontsource/roboto": "^4.5.8",
     "@mui/icons-material": "^5.11.16",
-    "@mui/material": "^5.12.1",
     "@mui/styled-engine-sc": "^5.12.0",
     "@mui/x-date-pickers": "^6.2.1",
     "@types/react-router-dom": "5.3.3",
