+++ src/components/datePicker/styled/index.tsx
