+++ src/components/inputOpt/index.tsx
