+++ src/components/tableOpt/index.tsx
@@ -1,272 +1,47 @@
 import * as React from 'react';
 import { useState, useEffect } from 'react';
 import axios from 'axios';
-import { alpha } from '@mui/material/styles';
 import Box from '@mui/material/Box';
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
 import TableContainer from '@mui/material/TableContainer';
-import TableHead from '@mui/material/TableHead';
 import TablePagination from '@mui/material/TablePagination';
 import TableRow from '@mui/material/TableRow';
-import TableSortLabel from '@mui/material/TableSortLabel';
-import Toolbar from '@mui/material/Toolbar';
-import Typography from '@mui/material/Typography';
 import Paper from '@mui/material/Paper';
 import Checkbox from '@mui/material/Checkbox';
-import IconButton from '@mui/material/IconButton';
-import Tooltip from '@mui/material/Tooltip';
 import FormControlLabel from '@mui/material/FormControlLabel';
+import FormControl from '@mui/material/FormControl';
 import Switch from '@mui/material/Switch';
-import DeleteIcon from '@mui/icons-material/Delete';
-import FilterListIcon from '@mui/icons-material/FilterList';
-import { visuallyHidden } from '@mui/utils';
-
+import { Order } from '../../constants';
+import { DataCanais } from '../../services';
+import { getComparator, stableSort } from '../../utils';
 import TextField from '@mui/material/TextField';
 import Stack from '@mui/material/Stack';
 import Autocomplete from '@mui/material/Autocomplete';
-
-interface Data {
-  CpfCnpj: string,
-  Telefone: string,
-  dataAtualizacao: string,
-  dataCriacao: string,
-  SistemaOrigem: string,
-  inOptInOut: string,
-}
-
-function createData(
-  CpfCnpj: string,
-  Telefone: string,
-  dataAtualizacao: string,
-  dataCriacao: string,
-  SistemaOrigem: string,
-  inOptInOut: string,
-): Data {
-  return {
-    CpfCnpj,
-    Telefone,
-    dataAtualizacao,
-    dataCriacao,
-    SistemaOrigem,
-    inOptInOut,
-  };
-}
-
-function descendingComparator<T>(a: T, b: T, orderBy: keyof T) {
-  if (b[orderBy] < a[orderBy]) {
-    return -1;
-  }
-  if (b[orderBy] > a[orderBy]) {
-    return 1;
-  }
-  return 0;
-}
-
-type Order = 'asc' | 'desc';
-
-function getComparator<Key extends keyof any>(
-  order: Order,
-  orderBy: Key,
-): (
-  a: { [key in Key]: number | string },
-  b: { [key in Key]: number | string },
-) => number {
-  return order === 'desc'
-    ? (a, b) => descendingComparator(a, b, orderBy)
-    : (a, b) => -descendingComparator(a, b, orderBy);
-}
-
-function stableSort<T>(array: readonly T[], comparator: (a: T, b: T) => number) {
-  const stabilizedThis = array.map((el, index) => [el, index] as [T, number]);
-  stabilizedThis.sort((a, b) => {
-    const order = comparator(a[0], b[0]);
-    if (order !== 0) {
-      return order;
-    }
-    return a[1] - b[1];
-  });
-  return stabilizedThis.map((el) => el[0]);
-}
-
-interface HeadCell {
-  disablePadding: boolean;
-  id: keyof Data;
-  label: string;
-  numeric: boolean;
-}
-
-const headCells: readonly HeadCell[] = [
-  {
-    id: 'CpfCnpj',
-    numeric: false,
-    disablePadding: true,
-    label: 'CPF/CNPJ',
-  },
-  {
-    id: 'Telefone',
-    numeric: true,
-    disablePadding: false,
-    label: 'Telefone',
-  },
-  {
-    id: 'dataAtualizacao',
-    numeric: true,
-    disablePadding: false,
-    label: 'Ultima Atualização',
-  },
-  {
-    id: 'dataCriacao',
-    numeric: true,
-    disablePadding: false,
-    label: 'Primeiro Consentimento',
-  },
-  {
-    id: 'SistemaOrigem',
-    numeric: true,
-    disablePadding: false,
-    label: 'Canal',
-  },
-  {
-    id: 'inOptInOut ',
-    numeric: true,
-    disablePadding: false,
-    label: 'Opção',
-  },
-];
-
-interface EnhancedTableProps {
-  numSelected: number;
-  onRequestSort: (event: React.MouseEvent<unknown>, property: keyof Data) => void;
-  onSelectAllClick: (event: React.ChangeEvent<HTMLInputElement>) => void;
-  order: Order;
-  orderBy: string;
-  rowCount: number;
-}
-
-function EnhancedTableHead(props: EnhancedTableProps) {
-  const { onSelectAllClick, order, orderBy, numSelected, rowCount, onRequestSort } =
-    props;
-  const createSortHandler =
-    (property: keyof Data) => (event: React.MouseEvent<unknown>) => {
-      onRequestSort(event, property);
-    };
-
-  return (
-    <TableHead>
-      <TableRow>
-        <TableCell padding="checkbox">
-          <Checkbox
-            color="primary"
-            indeterminate={numSelected > 0 && numSelected < rowCount}
-            checked={rowCount > 0 && numSelected === rowCount}
-            onChange={onSelectAllClick}
-            inputProps={{
-              'aria-label': 'select all desserts',
-            }}
-          />
-        </TableCell>
-        {headCells.map((headCell) => (
-          <TableCell
-            key={headCell.id}
-            align={headCell.numeric ? 'right' : 'left'}
-            padding={headCell.disablePadding ? 'none' : 'normal'}
-            sortDirection={orderBy === headCell.id ? order : false}
-          >
-            <TableSortLabel
-              active={orderBy === headCell.id}
-              direction={orderBy === headCell.id ? order : 'asc'}
-              onClick={createSortHandler(headCell.id)}
-            >
-              {headCell.label}
-              {orderBy === headCell.id ? (
-                <Box component="span" sx={visuallyHidden}>
-                  {order === 'desc' ? 'sorted descending' : 'sorted ascending'}
-                </Box>
-              ) : null}
-            </TableSortLabel>
-          </TableCell>
-        ))}
-      </TableRow>
-    </TableHead>
-  );
-}
-
-interface EnhancedTableToolbarProps {
-  numSelected: number;
-}
-
-function EnhancedTableToolbar(props: EnhancedTableToolbarProps) {
-  const { numSelected } = props;
-
-
-  return (
-    <Toolbar
-      sx={{
-        pl: { sm: 2 },
-        pr: { xs: 1, sm: 1 },
-        ...(numSelected > 0 && {
-          bgcolor: (theme) =>
-            alpha(theme.palette.primary.main, theme.palette.action.activatedOpacity),
-        }),
-      }}
-    >
-      {numSelected > 0 ? (
-        <Typography
-          sx={{ flex: '1 1 100%' }}
-          color="inherit"
-          variant="subtitle1"
-          component="div"
-        >
-          {numSelected} selected
-        </Typography>
-      ) : (
-        <Typography
-          sx={{ flex: '1 1 100%' }}
-          variant="h6"
-          id="tableTitle"
-          component="div"
-        >
-          
-        </Typography>
-      )}
-      {numSelected > 0 ? (
-        <Tooltip title="Delete">
-          <IconButton>
-            <DeleteIcon />
-          </IconButton>
-        </Tooltip>
-      ) : (
-        <Tooltip title="Filter list">
-          <IconButton>
-            <FilterListIcon />
-          </IconButton>
-        </Tooltip>
-      )}
-    </Toolbar>
-  );
-}
+import EnhancedTableHead from '../EnhancedTableHead';
+import EnhancedTableToolbar from '../EnhancedTableToolBar';
+import Grid2 from '@mui/material/Unstable_Grid2/Grid2';
+import CheckOpt from '../checkOpt';
 
 export default function EnhancedTable() {
   const [order, setOrder] = React.useState<Order>('asc');
-  const [orderBy, setOrderBy] = React.useState<keyof Data>('CpfCnpj');
+  const [orderBy, setOrderBy] = React.useState<keyof DataCanais>('CpfCnpj');
   const [selected, setSelected] = React.useState<readonly string[]>([]);
   const [page, setPage] = React.useState(0);
   const [dense, setDense] = React.useState(false);
   const [rowsPerPage, setRowsPerPage] = React.useState(5);
 
-  const [canais, setCanais] = useState<Data[]>([]);
+  const [canais, setCanais] = useState<DataCanais[]>([]);
 
   const fetchCanais = async () => {
     const response = await axios.get('http://localhost:3001/canais');
     setCanais(response.data);
-    // console.log(response.data);
   }
 
   const handleRequestSort = (
     event: React.MouseEvent<unknown>,
-    property: keyof Data,
+    property: keyof DataCanais,
   ) => {
     const isAsc = orderBy === property && order === 'asc';
     setOrder(isAsc ? 'desc' : 'asc');
@@ -337,7 +112,52 @@ export default function EnhancedTable() {
     }
     
   }
-  
+
+  const handleChangeInOut = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const elIn = event.target.ariaLabel === 'opt-in';
+    const elOut = event.target.ariaLabel === 'opt-out';
+    const elValue = event.target.checked;
+    // console.log("====================================");
+    // console.log('elin', elIn);
+    // console.log('elout', elOut);
+    // console.log('elvalue', elValue);
+    if(elIn && elValue){
+      const filteredRows = visibleRows.filter((row: any) => {
+        return row.inOptInOut === true;
+      })
+      setVisibleRows(filteredRows);
+      setCanais(filteredRows);
+    }else if(elOut && elValue){
+      const filteredRows = visibleRows.filter((row: any) => {
+        return row.inOptInOut === false;
+      })
+      setVisibleRows(filteredRows);
+      setCanais(filteredRows);
+    }else{
+      fetchCanais();
+    }
+  }
+
+  const handleCheckOut = (elIn: any, elOut:any, elVal:any) => {
+    
+    if(elIn && elVal){
+      const filteredRows = visibleRows.filter((row: any) => {
+        return row.inOptInOut === true;
+      })
+      setVisibleRows(filteredRows);
+      setCanais(filteredRows);
+    }else if(elOut && elVal){
+      const filteredRows = visibleRows.filter((row: any) => {
+        return row.inOptInOut === false;
+      })
+      setVisibleRows(filteredRows);
+      setCanais(filteredRows);
+    }else{
+      fetchCanais();
+    }
+
+  }
+
   useEffect(() => {
     fetchCanais();
   }, []);
@@ -349,13 +169,33 @@ export default function EnhancedTable() {
       page * rowsPerPage + rowsPerPage,
     ));
   }, [canais, order, orderBy, page, rowsPerPage]);
-
-
+  
   return (
     <Box sx={{ width: '100%' }}>
-      <Stack spacing={2} sx={{ width: 300 }}>
-        <TextField onChange={handleSearch} />
-      </Stack>
+
+      <Grid2 container spacing={2}>
+        
+        <Grid2 xs={6} md={4}>
+          <Stack spacing={2} sx={{ width: 400 }}>
+            <TextField onChange={handleSearch} />
+          </Stack>
+        </Grid2>
+
+        {/* <Grid2 xs={6} md={4}>
+          <FormControl variant="standard">
+            <FormControlLabel control={
+              <Checkbox
+              onChange={handleChangeInOut} 
+              inputProps={{'aria-label': 'opt-in'}} />} label="OPT IN" />
+            <FormControlLabel control={
+              <Checkbox
+              onChange={handleChangeInOut} 
+              inputProps={{'aria-label': 'opt-out'}} />} label="OPT OUT" />
+          </FormControl>
+        </Grid2> */}
+        <CheckOpt handleCheckOut={handleCheckOut} />
+      </Grid2>
+
       <Paper sx={{ width: '100%', mb: 2 }}>
         <EnhancedTableToolbar numSelected={selected.length} />
         <TableContainer>
